requirements.txt
Pillow>=10.0
ImageHash>=4.3